agregado Terraform

+++ 01-un-nodo-simple/provision/instala-ws.sh
@@ -13,20 +13,37 @@ export APT_OPTIONS_NEW=' -y --allow-downgrades --allow-remove-essential --allow-
 sudo apt-get install sshpass unzip ${APT_OPTIONS}
 
 instala_terraform() {
-  wget https://releases.hashicorp.com/terraform/1.0.2/terraform_1.0.2_linux_amd64.zip
-  unzip terraform_1.0.2_linux_amd64.zip 
-  sudo mv terraform /usr/local/bin/
+  VERSION=1.0.2
+  TF_ZIP_FILENAME=terraform_${VERSION}_linux_amd64.zip
+  TF_URL=https://releases.hashicorp.com/terraform/${VERSION}/${TF_ZIP_FILENAME}
+
+  cd /tmp
+  [ -f ${TF_ZIP_FILENAME} ] || wget ${TF_URL}
+  [ -f terraform ] || unzip ${TF_ZIP_FILENAME}
+  [ -f /usr/local/bin/terraform ] || sudo mv terraform /usr/local/bin/
   sudo chmod a+rx /usr/local/bin/terraform
 
   terraform version
 
-  terraform -install-autocomplete
+  grep "complete -C /usr/local/bin/terraform terraform" ~/.bashrc >/dev/null || terraform -install-autocomplete
 
 }
 
 configura_proxmox() {
-  export PM_PASS="admin"
-  export PM_USER="root@pam"
+  cd /vagrant/tf/
+
+  cat > .env <<!EOF
+export PM_PASS="admin"
+export PM_USER="root@pam"
+!EOF
 
   terraform init
 }
+
+instala_terraform
+configura_proxmox
+
+# genera claves SSH para acceder a los CT y VM
+[ -d /vagrant/tf/ssh-keys ] || mkdir /vagrant/tf/ssh-keys
+[ -f ~/.ssh/id_rsa-tf.pub ] || ( ssh-keygen -b 2048 -t rsa -f /vagrant/tf/ssh-keys/id_rsa-tf -q -N "" ; chmod og-rwx /vagrant/tf/ssh-keys/id_rsa-tf )
+
